vap/requirements.txt
# MCP SDK (install with: pip install mcp)
# mcp
pyyaml>=6.0
pydantic>=2.0.0
